sam4l: spi: update to new register interface
