--- chips/sam4l/src/regs/mod.rs
@@ -1,1 +1,0 @@
-pub mod spi;
