+++ chips/sam4l/src/spi.rs
@@ -12,7 +12,7 @@ use dma::DMAChannel;
 use dma::DMAClient;
 use dma::DMAPeripheral;
 use kernel::ReturnCode;
-use kernel::common::VolatileCell;
+use kernel::common::regs::{ReadOnly, ReadWrite, WriteOnly};
 use kernel::hil::spi;
 use kernel::hil::spi::ClockPhase;
 use kernel::hil::spi::ClockPolarity;
@@ -20,56 +20,136 @@ use kernel::hil::spi::SpiMasterClient;
 use kernel::hil::spi::SpiSlaveClient;
 use pm;
 
-/// The registers used to interface with the hardware
 #[repr(C)]
-struct SpiRegisters {
-    cr: VolatileCell<u32>,       //       0x0
-    mr: VolatileCell<u32>,       //       0x4
-    rdr: VolatileCell<u32>,      //      0x8
-    tdr: VolatileCell<u32>,      //      0xC
-    sr: VolatileCell<u32>,       //       0x10
-    ier: VolatileCell<u32>,      //      0x14
-    idr: VolatileCell<u32>,      //      0x18
-    imr: VolatileCell<u32>,      //      0x1C
-    _reserved0: [u32; 4],        //        0x20, 0x24, 0x28, 0x2C
-    csr0: VolatileCell<u32>,     //     0x30
-    csr1: VolatileCell<u32>,     //     0x34
-    csr2: VolatileCell<u32>,     //     0x38
-    csr3: VolatileCell<u32>,     //     0x3C
-    _reserved1: [u32; 41],       //       0x40 - 0xE0
-    wpcr: VolatileCell<u32>,     //     0xE4
-    wpsr: VolatileCell<u32>,     //     0xE8
-    _reserved2: [u32; 3],        //        0xEC - 0xF4
-    features: VolatileCell<u32>, // 0xF8
-    version: VolatileCell<u32>,  //  0xFC
+pub struct Registers {
+    pub cr: WriteOnly<u32, Control::Register>,
+    pub mr: ReadWrite<u32, Mode::Register>,
+    pub rdr: ReadOnly<u32>,
+    pub tdr: WriteOnly<u32, TransmitData::Register>,
+    pub sr: ReadOnly<u32, Status::Register>,
+    pub ier: WriteOnly<u32, InterruptFlags::Register>,
+    pub idr: WriteOnly<u32, InterruptFlags::Register>,
+    pub imr: ReadOnly<u32, InterruptFlags::Register>,
+    _reserved0: [ReadOnly<u32>; 4],
+    pub csr: [ReadWrite<u32, ChipSelectParams::Register>; 4],
+    _reserved1: [ReadOnly<u32>; 41],
+    pub wpcr: ReadWrite<u32, WriteProtectionControl::Register>,
+    pub wpsr: ReadOnly<u32>,
+    _reserved2: [ReadOnly<u32>; 3],
+    pub features: ReadOnly<u32>,
+    pub version: ReadOnly<u32>,
 }
+
+register_bitfields![u32,
+    Control [
+        LASTXFER 24,
+        FLUSHFIFO 8,
+        SWRST 7,
+        SPIDIS 1,
+        SPIEN 0
+    ],
+
+    /// Mode of the SPI peripheral.
+    Mode [
+        /// Delay between chip selects
+        DLYBCS   OFFSET(24)  NUMBITS(8) [],
+        /// Peripheral chip select
+        PCS      OFFSET(16)  NUMBITS(4) [
+            /// One-hot encoding
+            PCS0 = 0b1110,
+            PCS1 = 0b1101,
+            PCS2 = 0b1011,
+            PCS3 = 0b0111
+        ],
+        /// Local loopback enable
+        LLB      OFFSET( 7)  NUMBITS(1) [],
+        /// FIFO in reception enable
+        RXFIFOEN OFFSET( 6)  NUMBITS(1) [],
+        /// Mode fault detection
+        MODFDIS  OFFSET( 4)  NUMBITS(1) [],
+        /// Chip select decode
+        PCSDEC   OFFSET( 2)  NUMBITS(1) [],
+        /// Peripheral select
+        PS       OFFSET( 1)  NUMBITS(1) [],
+        /// Master/slave mode
+        MSTR     OFFSET( 0)  NUMBITS(1) []
+    ],
+
+    TransmitData [
+        LASTXFER OFFSET(24)  NUMBITS(1),
+        PCS      OFFSET(16)  NUMBITS(4),
+        TD       OFFSET(0)   NUMBITS(16)
+    ],
+
+    Status [
+        SPIENS  OFFSET(16),
+        UNDES   OFFSET(10),
+        TXEMPTY OFFSET(9),
+        NSSR    OFFSET(8),
+        OVRES   OFFSET(3),
+        MODF    OFFSET(2),
+        TDRE    OFFSET(1),
+        RDRF    OFFSET(0)
+    ],
+
+    InterruptFlags [
+        UNDES 10,
+        TXEMPTY 9,
+        NSSR 8,
+        OVRES 3,
+        MODF 2,
+        TDRE 1,
+        RDRF 0
+    ],
+
+    ChipSelectParams [
+        DLYBCT OFFSET(24)  NUMBITS(8) [],
+        DLYBS  OFFSET(16)  NUMBITS(8) [],
+        SCBR   OFFSET(8)   NUMBITS(8) [],
+        BITS   OFFSET(4)   NUMBITS(8) [
+            Eight = 0,
+            Nine = 1,
+            Ten = 2,
+            Eleven = 3,
+            Twelve = 4,
+            Thirteen = 5,
+            Fourteen = 6,
+            Fifteen = 7,
+            Sixteen = 8,
+            Four = 9,
+            Five = 10,
+            Six = 11,
+            Seven = 12
+        ],
+        CSAAT OFFSET(3)  NUMBITS(1) [
+            ActiveAfterTransfer = 1,
+            InactiveAfterTransfer = 0
+        ],
+        CSNAAT OFFSET(2)  NUMBITS(1) [
+            DoNotRiseBetweenTransfers = 0,
+            RiseBetweenTransfers = 1
+        ],
+        NCPHA OFFSET(1)  NUMBITS(1) [
+            CaptureLeading = 1,
+            CaptureTrailing = 0
+        ],
+        CPOL OFFSET(0)  NUMBITS(1) [
+            InactiveHigh = 1,
+            InactiveLow = 0
+        ]
+    ],
+
+    WriteProtectionControl [
+        SPIWPKEY OFFSET(8) NUMBITS(24) [
+            Key = 0x535049
+        ],
+        SPIWPEN OFFSET(0) NUMBITS(1) []
+    ]
+];
 
 #[allow(unused_variables, dead_code)]
 // Per-register masks defined in the SPI manual in chapter 26.8
 mod spi_consts {
-    pub mod cr {
-        pub const SPIEN: u32 = 1 << 0;
-        pub const SPIDIS: u32 = 1 << 1;
-        pub const SWRST: u32 = 1 << 7;
-        pub const FLUSHFIFO: u32 = 1 << 8;
-        pub const LASTXFER: u32 = 1 << 24;
-    }
-
-    pub mod mr {
-        pub const MSTR: u32 = 1 << 0;
-        pub const PS: u32 = 1 << 1;
-        pub const PCSDEC: u32 = 1 << 2;
-        pub const MODFDIS: u32 = 1 << 4;
-        pub const RXFIFOEN: u32 = 1 << 6;
-        pub const LLB: u32 = 1 << 7;
-        pub const PCS_MASK: u32 = 0b1111 << 16;
-        pub const PCS0: u32 = 0b1110 << 16;
-        pub const PCS1: u32 = 0b1101 << 16;
-        pub const PCS2: u32 = 0b1011 << 16;
-        pub const PCS3: u32 = 0b0111 << 16;
-        pub const DLYBCS_MASK: u32 = 0xFF << 24;
-    }
-
     pub mod rdr {
         pub const RD: u32 = 0xFFFF;
     }
@@ -78,45 +158,6 @@ mod spi_consts {
         pub const TD: u32 = 0xFFFF;
         // PCSx masks from MR also apply here
         // LASTXFER from CR also applies here
-    }
-
-    pub mod sr {
-        // These same bits are used in IDR, IER, and IMR.
-        pub const RDRF: u32 = 1 << 0;
-        pub const TDRE: u32 = 1 << 1;
-        pub const MODF: u32 = 1 << 2;
-        pub const OVRES: u32 = 1 << 3;
-        pub const NSSR: u32 = 1 << 8;
-        pub const TXEMPTY: u32 = 1 << 9;
-        pub const UNDES: u32 = 1 << 10;
-
-        // This only exists in the SR
-        pub const SPIENS: u32 = 1 << 16;
-    }
-
-    // These bit masks apply to CSR0; CSR1, CSR2, CSR3
-    pub mod csr {
-        pub const CPOL: u32 = 1 << 0;
-        pub const NCPHA: u32 = 1 << 1;
-        pub const CSNAAT: u32 = 1 << 2;
-        pub const CSAAT: u32 = 1 << 3;
-        pub const BITS_MASK: u32 = 0x1111 << 4;
-        pub const BITS8: u32 = 0b0000 << 4;
-        pub const BITS9: u32 = 0b0001 << 4;
-        pub const BITS10: u32 = 0b0010 << 4;
-        pub const BITS11: u32 = 0b0011 << 4;
-        pub const BITS12: u32 = 0b0100 << 4;
-        pub const BITS13: u32 = 0b0101 << 4;
-        pub const BITS14: u32 = 0b0110 << 4;
-        pub const BITS15: u32 = 0b0111 << 4;
-        pub const BITS16: u32 = 0b1000 << 4;
-        pub const BITS4: u32 = 0b1001 << 4;
-        pub const BITS5: u32 = 0b1010 << 4;
-        pub const BITS6: u32 = 0b1011 << 4;
-        pub const BITS7: u32 = 0b1100 << 4;
-        pub const SCBR_MASK: u32 = 0xFF << 8;
-        pub const DLYBS_MASK: u32 = 0xFF << 16;
-        pub const DLYBCT_MASK: u32 = 0xFF << 24;
     }
 }
 
@@ -139,7 +180,7 @@ pub enum SpiRole {
 
 /// The SAM4L supports four peripherals.
 pub struct Spi {
-    registers: *mut SpiRegisters,
+    registers: *mut Registers,
     client: Cell<Option<&'static SpiMasterClient>>,
     dma_read: Cell<Option<&'static DMAChannel>>,
     dma_write: Cell<Option<&'static DMAChannel>>,
@@ -159,7 +200,7 @@ impl Spi {
     /// Creates a new SPI object, with peripheral 0 selected
     pub const fn new() -> Spi {
         Spi {
-            registers: SPI_BASE as *mut SpiRegisters,
+            registers: SPI_BASE as *mut Registers,
             client: Cell::new(None),
             dma_read: Cell::new(None),
             dma_write: Cell::new(None),
@@ -172,54 +213,52 @@ impl Spi {
     }
 
     fn init_as_role(&self, role: SpiRole) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         self.role.set(role);
         self.enable_clock();
 
         if self.role.get() == SpiRole::SpiMaster {
             // Only need to set LASTXFER if we are master
-            regs.cr.set(spi_consts::cr::LASTXFER);
+            regs.cr.write(Control::LASTXFER::SET);
         }
 
         // Sets bits per transfer to 8
         let mut csr = self.read_active_csr();
-        csr &= !spi_consts::csr::BITS_MASK;
-        csr |= spi_consts::csr::BITS8;
+        csr &= !(0x1111 << 4);
+        csr |= 0b0000 << 4;
         self.write_active_csr(csr);
 
         // Set mode to master or slave
-        let mut mode = regs.mr.get();
-        match self.role.get() {
-            SpiRole::SpiMaster => mode |= spi_consts::mr::MSTR,
-            SpiRole::SpiSlave => mode &= !spi_consts::mr::MSTR,
-        }
+        let mode = match self.role.get() {
+            SpiRole::SpiMaster => Mode::MSTR::SET,
+            SpiRole::SpiSlave => Mode::MSTR::CLEAR,
+        };
 
         // Disable mode fault detection (open drain outputs not supported)
-        mode |= spi_consts::mr::MODFDIS;
-        regs.mr.set(mode);
+        regs.mr.modify(mode + Mode::MODFDIS::SET);
     }
 
     pub fn enable(&self) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         self.enable_clock();
-        regs.cr.set(spi_consts::cr::SPIEN);
+        regs.cr.write(Control::SPIEN::SET);
 
         if self.role.get() == SpiRole::SpiSlave {
-            regs.ier.set(spi_consts::sr::NSSR); // Enable NSSR
+            regs.ier.write(InterruptFlags::NSSR::SET); // Enable NSSR
         }
     }
 
     pub fn disable(&self) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         self.dma_read.get().map(|read| read.disable());
         self.dma_write.get().map(|write| write.disable());
-        regs.cr.set(spi_consts::cr::SPIDIS);
+        regs.cr.write(Control::SPIDIS::SET);
 
         if self.role.get() == SpiRole::SpiSlave {
-            regs.idr.set(spi_consts::sr::NSSR); // Disable NSSR
+            regs.idr.write(InterruptFlags::NSSR::SET);; // Disable NSSR
         }
     }
 
@@ -255,29 +294,29 @@ impl Spi {
             scbr += 1;
         }
         let mut csr = self.read_active_csr();
-        csr = (csr & !spi_consts::csr::SCBR_MASK) | ((scbr & 0xFF) << 8);
+        csr = (csr & !(0xFF << 8)) | ((scbr & 0xFF) << 8);
         self.write_active_csr(csr);
         clock / scbr
     }
 
     pub fn get_baud_rate(&self) -> u32 {
         let clock = 48000000;
-        let scbr = (self.read_active_csr() & spi_consts::csr::SCBR_MASK) >> 8;
+        let scbr = (self.read_active_csr() & (0xFF << 8)) >> 8;
         clock / scbr
     }
 
     fn set_clock(&self, polarity: ClockPolarity) {
         let mut csr = self.read_active_csr();
         match polarity {
-            ClockPolarity::IdleHigh => csr |= spi_consts::csr::CPOL,
-            ClockPolarity::IdleLow => csr &= !spi_consts::csr::CPOL,
+            ClockPolarity::IdleHigh => csr |= 0x01,
+            ClockPolarity::IdleLow => csr &= !0x01,
         };
         self.write_active_csr(csr);
     }
 
     fn get_clock(&self) -> ClockPolarity {
         let csr = self.read_active_csr();
-        let polarity = csr & spi_consts::csr::CPOL;
+        let polarity = csr & 0x01;
         match polarity {
             0 => ClockPolarity::IdleLow,
             _ => ClockPolarity::IdleHigh,
@@ -287,15 +326,15 @@ impl Spi {
     fn set_phase(&self, phase: ClockPhase) {
         let mut csr = self.read_active_csr();
         match phase {
-            ClockPhase::SampleLeading => csr |= spi_consts::csr::NCPHA,
-            ClockPhase::SampleTrailing => csr &= !spi_consts::csr::NCPHA,
+            ClockPhase::SampleLeading => csr |= 0x02,
+            ClockPhase::SampleTrailing => csr &= !0x02,
         };
         self.write_active_csr(csr);
     }
 
     fn get_phase(&self) -> ClockPhase {
         let csr = self.read_active_csr();
-        let phase = csr & spi_consts::csr::NCPHA;
+        let phase = csr & 0x02;
         match phase {
             0 => ClockPhase::SampleTrailing,
             _ => ClockPhase::SampleLeading,
@@ -305,32 +344,28 @@ impl Spi {
     pub fn set_active_peripheral(&self, peripheral: Peripheral) {
         // Slave cannot set active peripheral
         if self.role.get() == SpiRole::SpiMaster {
-            let regs: &SpiRegisters = unsafe { &*self.registers };
-            let peripheral_number: u32 = match peripheral {
-                Peripheral::Peripheral0 => spi_consts::mr::PCS0,
-                Peripheral::Peripheral1 => spi_consts::mr::PCS1,
-                Peripheral::Peripheral2 => spi_consts::mr::PCS2,
-                Peripheral::Peripheral3 => spi_consts::mr::PCS3,
+            let regs: &Registers = unsafe { &*self.registers };
+            let mr = match peripheral {
+                Peripheral::Peripheral0 => Mode::PCS::PCS0,
+                Peripheral::Peripheral1 => Mode::PCS::PCS1,
+                Peripheral::Peripheral2 => Mode::PCS::PCS2,
+                Peripheral::Peripheral3 => Mode::PCS::PCS3,
             };
-            let mut mr = regs.mr.get();
-            mr = (mr & !spi_consts::mr::PCS_MASK) | peripheral_number;
-            regs.mr.set(mr);
+            regs.mr.modify(mr);
         }
     }
 
     /// Returns the currently active peripheral
     pub fn get_active_peripheral(&self) -> Peripheral {
         if self.role.get() == SpiRole::SpiMaster {
-            let regs: &SpiRegisters = unsafe { &*self.registers };
+            let regs: &Registers = unsafe { &*self.registers };
 
-            let mr = regs.mr.get();
-            let peripheral_number = mr & (spi_consts::mr::PCS_MASK);
-
+            let peripheral_number = regs.mr.read(Mode::PCS);
             match peripheral_number {
-                spi_consts::mr::PCS0 => Peripheral::Peripheral0,
-                spi_consts::mr::PCS1 => Peripheral::Peripheral1,
-                spi_consts::mr::PCS2 => Peripheral::Peripheral2,
-                spi_consts::mr::PCS3 => Peripheral::Peripheral3,
+                0b1110 => Peripheral::Peripheral0,
+                0b1101 => Peripheral::Peripheral1,
+                0b1011 => Peripheral::Peripheral2,
+                0b0111 => Peripheral::Peripheral3,
                 _ => {
                     // Invalid configuration
                     Peripheral::Peripheral0
@@ -345,25 +380,25 @@ impl Spi {
     /// Returns the value of CSR0, CSR1, CSR2, or CSR3,
     /// whichever corresponds to the active peripheral
     fn read_active_csr(&self) -> u32 {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         match self.get_active_peripheral() {
-            Peripheral::Peripheral0 => regs.csr0.get(),
-            Peripheral::Peripheral1 => regs.csr1.get(),
-            Peripheral::Peripheral2 => regs.csr2.get(),
-            Peripheral::Peripheral3 => regs.csr3.get(),
+            Peripheral::Peripheral0 => regs.csr[0].get(),
+            Peripheral::Peripheral1 => regs.csr[1].get(),
+            Peripheral::Peripheral2 => regs.csr[2].get(),
+            Peripheral::Peripheral3 => regs.csr[3].get(),
         }
     }
     /// Sets the Chip Select Register (CSR) of the active peripheral
     /// (CSR0, CSR1, CSR2, or CSR3).
     fn write_active_csr(&self, value: u32) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         match self.get_active_peripheral() {
-            Peripheral::Peripheral0 => regs.csr0.set(value),
-            Peripheral::Peripheral1 => regs.csr1.set(value),
-            Peripheral::Peripheral2 => regs.csr2.set(value),
-            Peripheral::Peripheral3 => regs.csr3.set(value),
+            Peripheral::Peripheral0 => regs.csr[0].set(value),
+            Peripheral::Peripheral1 => regs.csr[1].set(value),
+            Peripheral::Peripheral2 => regs.csr[2].set(value),
+            Peripheral::Peripheral3 => regs.csr[3].set(value),
         };
     }
 
@@ -380,11 +415,10 @@ impl Spi {
     }
 
     pub fn handle_interrupt(&self) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
-        let sr = regs.sr.get();
+        let regs: &Registers = unsafe { &*self.registers };
 
         self.slave_client.get().map(|client| {
-            if (sr & spi_consts::sr::NSSR) != 0 {
+            if regs.sr.is_set(Status::NSSR) {
                 // NSSR
                 client.chip_selected()
             }
@@ -470,12 +504,12 @@ impl spi::SpiMaster for Spi {
     /// Write a byte to the SPI and discard the read; if an
     /// asynchronous operation is outstanding, do nothing.
     fn write_byte(&self, out_byte: u8) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         let tdr = (out_byte as u32) & spi_consts::tdr::TD;
         // Wait for data to leave TDR and enter serializer, so TDR is free
         // for this next byte
-        while (regs.sr.get() & spi_consts::sr::TDRE) == 0 {}
+        while !regs.sr.is_set(Status::TDRE) {}
         regs.tdr.set(tdr);
     }
 
@@ -488,11 +522,11 @@ impl spi::SpiMaster for Spi {
     /// Write a byte to the SPI and return the read; if an
     /// asynchronous operation is outstanding, do nothing.
     fn read_write_byte(&self, val: u8) -> u8 {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
 
         self.write_byte(val);
         // Wait for receive data register full
-        while (regs.sr.get() & spi_consts::sr::RDRF) == 0 {}
+        while !regs.sr.is_set(Status::RDRF) {}
         // Return read value
         regs.rdr.get() as u8
     }
@@ -548,13 +582,13 @@ impl spi::SpiMaster for Spi {
 
     fn hold_low(&self) {
         let mut csr = self.read_active_csr();
-        csr |= spi_consts::csr::CSAAT;
+        csr |= 1 << 3;
         self.write_active_csr(csr);
     }
 
     fn release_low(&self) {
         let mut csr = self.read_active_csr();
-        csr &= !spi_consts::csr::CSAAT;
+        csr &= !(1 << 3);
         self.write_active_csr(csr);
     }
 
@@ -587,7 +621,7 @@ impl spi::SpiSlave for Spi {
     /// This sets the value in the TDR register, to be sent as soon as the
     /// chip select pin is low.
     fn set_write_byte(&self, write_byte: u8) {
-        let regs: &SpiRegisters = unsafe { &*self.registers };
+        let regs: &Registers = unsafe { &*self.registers };
         regs.tdr.set(write_byte as u32);
     }
